requirements.txt
# ─────────────────────────────────────────────────────────────────
# ScholarAI — Python Dependencies
# Install with: pip install -r requirements.txt
# For Google Colab: pip install -r requirements.txt -q
# ─────────────────────────────────────────────────────────────────

# ── Web Framework ────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6       # File upload support for FastAPI

# ── ML Core ─────────────────────────────────────────────────────
torch>=2.0.0
transformers>=4.35.0           # BART, T5 models
sentence-transformers>=2.2.2   # Sentence-BERT
datasets>=2.14.0               # Optional: for training data
accelerate>=0.24.0             # HuggingFace training helper
scikit-learn>=1.3.0            # TF-IDF, cosine similarity, PCA

# ── NLP ──────────────────────────────────────────────────────────
nltk>=3.8.1
spacy>=3.7.0
# After installing spacy run: python -m spacy download en_core_web_sm

# ── Document Parsing ─────────────────────────────────────────────
pymupdf>=1.23.0                # PyMuPDF (fitz) — fast PDF extraction
pdfplumber>=0.10.0             # Alternative PDF extractor
PyPDF2>=3.0.0                  # Fallback PDF extractor
python-pptx>=0.6.21            # PowerPoint extraction

# ── Audio (TTS) ──────────────────────────────────────────────────
gTTS>=2.4.0                    # Google Text-to-Speech (online)
pyttsx3>=2.90                  # Offline TTS fallback
pydub>=0.25.1                  # Audio processing / conversion
# Optional (high quality offline TTS):
# TTS>=0.22.0                  # Coqui TTS (large download ~500MB)

# ── Video Generation ─────────────────────────────────────────────
moviepy>=1.0.3
Pillow>=10.0.0                 # PIL image rendering

# ── Security / Encryption ────────────────────────────────────────
cryptography>=41.0.0           # Fernet (AES-128 + HMAC)
# pycryptodome>=3.19.0         # Alternative AES (CBC mode)

# ── Utilities ────────────────────────────────────────────────────
numpy>=1.24.0
python-dotenv>=1.0.0           # Environment variable management
aiofiles>=23.2.0               # Async file I/O
httpx>=0.25.0                  # HTTP client for testing
pydantic>=2.4.0                # Data validation
